refactor : ValidatorTest 분리하기
